+++ src/main/java/xilef11/mc/runesofwizardry_classics/runes/RuneSpiritTools.java
@@ -56,4 +56,12 @@ public class RuneSpiritTools extends ClassicRune {
 			Set<BlockPos> dusts, TileEntityDustActive entity) {
 		return new RuneEntityUnimplemented(actualPattern, front, dusts, entity, this);
 	}
+	/* (non-Javadoc)
+	 * @see com.zpig333.runesofwizardry.api.IRune#getShortDesc()
+	 */
+	@Override
+	public String getShortDesc() {
+		//Decorative/WIP
+		return Refs.Lang.DEC;
+	}
 }