+++ changelog.md
@@ -2,5 +2,7 @@
 - Added config for the Rune of Locked Time stopping falling blocks
 - Added config blacklist for rune of level earth
 - Added config whitelist for farmland replaceable blocks 
+- Improved WIP runes documentation
+- Added config option to disable WIP runes
 
 Uses Runes of Wizardry 0.6.5