+++ changelog.md
@@ -4,5 +4,6 @@
 - Added config whitelist for farmland replaceable blocks 
 - Improved WIP runes documentation
 - Added config option to disable WIP runes
+- Made Rune of Rebirth ignore players and armour stands
 
 Uses Runes of Wizardry 0.6.5